backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.9.0",
    "mcp[cli]>=1.6.0",
    "pypdf>=4.2.0",
    "qdrant-client>=1.9.0",
    "numpy>=1.26.4",
    "qdrant-haystack>=9.1.1",
    "fastembed-haystack>=1.4.1",
    "pydantic-ai>=0.0.53",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "nest-asyncio>=1.6.0",
    "pre-commit>=4.2.0",
    "ruff>=0.11.2",
]
